refactor(client): fix imports messed up during rebase
